% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleNN.R
\name{class.ind}
\alias{class.ind}
\title{Convert factor to a dummy/indicator matrix}
\usage{
class.ind(cl)
}
\arguments{
\item{cl}{factor vector}
}
\value{
Appropriate indicator matrix.
}
\description{
\code{class.ind} take a factor vector of length $n$ with $k$ levels and
                 outputs an $n$ x $k$ indicator matrix.
}
\examples{
# Random factor data
dat <- factor(sample(1:3,100, replace=TRUE))
# Get the ggplot object
indicatorMatrix <- class.ind(dat)

}
