% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadAndLoad.R
\name{download_mnist}
\alias{download_mnist}
\title{Download the mnist data}
\usage{
download_mnist(where = "./")
}
\arguments{
\item{where}{Path to where the dataset should be saved. Default is current
working directory.}
}
\value{
Returns \code{TRUE} if successful.
}
\description{
\code{download_mnist} downloads the mnist dataset into \code{./mnist}.
}
\details{
The function creates the \code{mnist} directory in the specified directory \code{where},
downloads the dataset and unzips it. Use the function \code{\link{load_mnist}}
afterwards to load the dataset into the workspace. The dataset is downloaded
from \url{http://yann.lecun.com/exdb/mnist/}. Check out the webpage for
performance of various classifiers.
}
\examples{

\dontrun{
download_mnist('somePathToStore/')
load_mnist('samePathAsAbove/')
}
}

