% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadAndLoad.R
\name{load_mnist}
\alias{load_mnist}
\title{Load the mnist dataset into the workspace}
\usage{
load_mnist(where = "./")
}
\arguments{
\item{where}{Path to the folder containing the mnist directory. Default assumes
that the current working directory contains the mnist folder.}
}
\value{
Returns \code{TRUE} if successful.
}
\description{
\code{load_mnist} Loads the mnist dataset as the lists trainSet and testSet into
the current workspace.
}
\details{
Assuming that you have downloaded the dataset with the \code{\link{download_mnist}}
function, the \code{load_mnist} function loads the data into the lists trainSet and testSet.
Both lists have three elements named n, x and y. n is the number of samples, x is the
28 by 28 images in rows and y is the label of the images. This function is inspired by
the gist: brendan o'connor - gist.github.com/39760 - anyall.org
}
\examples{

\dontrun{
download_mnist('somePathToStore/')
load_mnist('samePathAsAbove/')
# Check the number of elements in training set
trainSet$n
# Plot a random digit
(ggDigit(trainSet$x[sample(1:trainSet$n,1),]))
}
}
