% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleNN.R
\name{matrixInLayer}
\alias{matrixInLayer}
\title{Matrix closure for neural network layers}
\usage{
matrixInLayer(init = FALSE, rS, cS, initPos = FALSE, initScale = 100)
}
\arguments{
\item{init}{boolean for whether we need to initialize values.}

\item{rS}{number of rows.}

\item{cS}{number of columns.}

\item{initPos}{boolean to indicate whether values need to be initialized as positive.}

\item{initScale}{scalar to truncate initialized values closer to zero.}
}
\value{
environment with the functions \code{setter} and \code{getter}
}
\description{
\code{matrixInLayer} getter and setter functions for a matrix
}
\examples{
testFun <- matrixInLayer(TRUE,10,10,TRUE,10)

}
\keyword{internal}
