% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnistr.R
\docType{package}
\name{mnistr}
\alias{mnistr}
\alias{mnistr-package}
\title{mnistr: A package to work with mnist data in R.}
\description{
The mnistr package provides functions to download, load
and plot the mnist dataset. Some examples are provided
on how to perform classification on the dataset.
}
\section{Explain categories of functions}{

The functions ...
}

