% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleNN.R
\name{reLU}
\alias{reLU}
\title{Rectified Linear Unit activation functon}
\usage{
reLU(X, deriv = FALSE)
}
\arguments{
\item{X}{numeric scalar, vector or matrix}

\item{deriv}{boolean indicating whether we should evaluate the function
or the derivative at the input.}
}
\value{
Same format as the input \code{X}.
}
\description{
\code{sigmoid} can take a scalar, vector or matrix and output
               elementwise application of the reLU nonlinearity/
               squashing function. There is an additional boolean
               flag for calculating the derivative. The deriv argument
               is needed for activation functions.
}
\examples{
reLU(-5)
reLU(0)
reLU(5)
reLU(1,deriv=TRUE)

}
