% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleNN.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid activation functon}
\usage{
sigmoid(X, deriv = FALSE)
}
\arguments{
\item{X}{numeric scalar, vector or matrix}

\item{deriv}{boolean indicating whether we should evaluate the function
or the derivative at the input.}
}
\value{
Same format as the input \code{X}.
}
\description{
\code{sigmoid} can take a scalar, vector or matrix and output
               elementwise application of the sigmoid nonlinearity/
               squashing function. There is an additional boolean
               flag for calculating the derivative. The deriv argument
               is needed for activation functions.
}
\examples{
sigmoid(-5)
sigmoid(0)
sigmoid(5)
sigmoid(0,deriv=TRUE)

}
