% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleNN.R
\name{softmax}
\alias{softmax}
\title{Softmax activation functon}
\usage{
softmax(X)
}
\arguments{
\item{X}{numeric vector or matrix (for minibatches)}
}
\value{
Same format as the input \code{X}.
}
\description{
\code{sigmoid} takes a vector and applies a softmax to it for 
               transforming it into probabilities. This is 
               normally used for an output layer in a neural
               network for doing classification.
}
\examples{
softmax(matrix(1:10,5,2))
softmax(matrix(1:10,2,5))

}
